PYIC-6089: change ReverificationResponse to record instead of class and move to common-services.domain
